backend/requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0

# AI & LangChain
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
openai==1.6.1
tiktoken==0.5.2
assemblyai==0.23.0

# PDF Processing
PyPDF2==3.0.1
reportlab==4.0.8

# Validation & Serialization
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Google Services
gspread==5.12.0
google-auth==2.25.2
google-api-python-client==2.111.0

# Notifications
aiosmtplib==3.0.1
jinja2==3.1.2

# Background Tasks
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
